+++ src/demo/index.js
@@ -1,4 +1,5 @@
-import { renderApp } from '../../src/app/app';
+import 'whatwg-fetch';
+import { renderApp } from '../app/app';
 
 const urlParams = new URLSearchParams(window.location.search);
 if (urlParams.has('small')) {