+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 0.0.25 - In progress
 ### Added
+- Removed nwb, added custom webpack configuration.
 - Added a theme option in the dataset config objects, with support for "light" and "dark".
+
 
 ## [0.0.24](https://www.npmjs.com/package/vitessce/v/0.0.24) - 2020-03-02
 ### Added